Refactor: SelectToggle 컴포넌트 모듈화 및 코드 정리
